repeat function

+++ src/service/deleteTask.js
@@ -1,16 +1,28 @@
-import { doc, deleteDoc } from 'firebase/firestore';
+import { doc, getDoc, setDoc, deleteDoc, serverTimestamp } from 'firebase/firestore';
 import { firestore } from './firebase'; // Import your Firestore instance
 
 export default async function HandleDeleteTask(id) {
+
   try {
     // Define the Firestore document reference for the task with the given ID
     const taskDocRef = doc(firestore, 'todos', id);
 
-    // Delete the document
-    await deleteDoc(taskDocRef);
+    // Get the task document data
+    const taskDocSnapshot = await getDoc(taskDocRef);
 
-    console.log('Task deleted successfully.');
+    if (taskDocSnapshot.exists()) {
+      // Check if the 'doesRepeat' field is set to 'on'
+      const doesRepeat = taskDocSnapshot.data().doesRepeat;
+
+      if (doesRepeat === 'on') {
+        // Update the 'date' field with the current date (or your desired logic)
+        await setDoc(taskDocRef, { done: serverTimestamp() }, { merge: true });
+      } else {
+        // Delete the document
+        await deleteDoc(taskDocRef);
+      }
+    }
   } catch (error) {
-    console.error('Error deleting task:', error);
+    console.error('Error deleting or updating task:', error);
   }
 }
